refactor(issue-2216): extract chatbot knowledge base to config file
